Final changes
Adding a new top-level mechanism, 'disrupts membrane potential'.  Not sure if this fits better somewhere else.  Getting rid of the terms that used to live under that term 'post-synaptic membrane potential modulator', and 'action potential propagation modulator'.   Then, getting rid of the top-level terms 'process or component of pathogenicity' and 'determinant of pathogenicity', leaving 'mechanism of pathogenicity' as the new top most term.